correction des erreurs énoncé lors de l'evaluation de Sprint
création de variables pour les coordonnées de joueur
séparation du modèle et de la console avec la méthode unTour
conversion des coordonnés des personnages en double

+++ src/application/modele/Joueur.java
@@ -30,7 +30,7 @@ public class Joueur extends Personnage{
 
 			@Override
 			public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
-				versDroite=((int)oldValue<(int)newValue);
+				versDroite=((double)oldValue<(double)newValue);
 			}	
 		});
 	}
